Getting APP to work using db.
Slowly converting the old code to work with the new database setup. Loging in and fetching events works fine.

+++ app/src/main/java/com/example/churro/meetly/MeetlyServerImpl.java
@@ -67,8 +67,8 @@ public class MeetlyServerImpl implements MeetlyServer {
             CallableStatement createStmt = con.prepareCall(create);
             createStmt.setInt(1, userToken);
             createStmt.setString(2, title);
-            createStmt.setInt(3, (int) startTime.getTimeInMillis());
-            createStmt.setInt(4, (int)endTime.getTimeInMillis());
+            createStmt.setLong(3, startTime.getTimeInMillis());
+            createStmt.setLong(4, endTime.getTimeInMillis());
             createStmt.setDouble(5, latitude);
             createStmt.setDouble(6, longitude);
             createStmt.registerOutParameter(7, Types.INTEGER);
@@ -98,8 +98,8 @@ public class MeetlyServerImpl implements MeetlyServer {
             modifyStmt.setInt(1, eventID);
             modifyStmt.setInt(2, userToken);
             modifyStmt.setString(3, title);
-            modifyStmt.setInt(4, (int) startTime.getTimeInMillis());
-            modifyStmt.setInt(5, (int) endTime.getTimeInMillis());
+            modifyStmt.setLong(4, startTime.getTimeInMillis());
+            modifyStmt.setLong(5, endTime.getTimeInMillis());
             modifyStmt.setDouble(6, latitude);
             modifyStmt.setDouble(7, longitude);
             modifyStmt.registerOutParameter(8, Types.INTEGER);
@@ -136,9 +136,9 @@ public class MeetlyServerImpl implements MeetlyServer {
                 event.title      = results.getString(4);
 
                 event.startTime  = Calendar.getInstance();
-                event.startTime.setTimeInMillis(results.getInt(5));
+                event.startTime.setTimeInMillis(results.getLong(5));
                 event.endTime    = Calendar.getInstance();
-                event.endTime.setTimeInMillis(results.getInt(6));
+                event.endTime.setTimeInMillis(results.getLong(6));
 
                 event.latitude   = results.getDouble(7);
                 event.longitude  = results.getDouble(8);
